regulator: qcom_smd: Implement support for voltages on BoB legs
The BoB regulator in at least PM660L has more than one output leg,
on which is possible to set the voltage.
Of course the BoB legs are powered by the BoB regulator with their
same resource ID, so they are effectively children of that vreg.

Implement the support for this kind of vregs: this is done by
sending a RPM message with a different key (in this case, the
keys that are referenced downstream are "pcv1", "pcv2", "pcv3")
with the voltage to set in microvolts, exactly as it is already
being done in the "uv" key case.

While at it, also add them to the PM660L regulators, as that's
what has been used as the PMIC of reference.

Signed-off-by: AngeloGioacchino Del Regno <kholk11@gmail.com>