arm64: dts: qcom: sdm660-xiaomi-clover: Fix regulator probe issues
For always-on or boot-on regulators their supplies must be specified before
them, because regulator core framework tries to resolve their supplies early
and fails with probe deferral error if it can't.

The supply for PM660L's l3b is buck-or-boost vreg_bob, but it is not
registered yet (it is described as last item). Move it upper above l3b
and others.

Also bob_pin1, bob_pin2 and bob_pin3 are not supported by driver,
remove them.

Signed-off-by: Alexey Minnekhanov <alexeymin@postmarketos.org>