OPP: Remove the unused argument to config_clks_t
The OPP core needs to take care of a special case, where the OPPs aren't
available for a device, but in order to keep the same unified interface
for the driver, the same OPP core API must take care of performing a
simple clk_set_rate() for the device.

This required the extra argument, but that is used only within the OPP
core and the drivers don't need to take care of that.

Simplify the external API and handle it differently within the OPP core.

This shouldn't result in any functional change.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>